DRT-4938 - Using the Peek-Lock Strategy
2.5.1 Peek-Lock
This strategy locks a message in the queue and sends the message to the receiver with an identifier for the lock. Whilst the message is locked, it cannot be seen by any other operations – it is essentially invisible to all but the holder of the lock. The holder of the lock is expected to process the message and then either remove the message from the queue, or release the lock to return the message to the queue.
The lock has an expiry time, after which it is automatically released. This protects the message being “lost” if the receiver is suddenly halted. The receiver can renew the lock on the message to prevent it being released prematurely.
The Peek-Lock strategy produces at-least-once messaging; messages will be delivered and processed at least once, possibly multiple times. A message is only removed from the queue once it has been processed to completion, but errors in communication may cause it to be delivered multiple times.
Due to the requirement to de-duplicate messages received from multiple AME instances, having a message delivered multiple times should not cause any significant impact on processing. As such, the Peek-Lock strategy is the recommended strategy.